booking review 구현 완료 및 delete와 review 실행 시 게스트가 예약한 booking에 한해서만 실행될 수 있게 수정

+++ src/main/java/org/example/controller/BookingController.java
@@ -7,6 +7,7 @@ import org.example.dto.Room;
 import org.example.service.BookingService;
 import org.example.util.Util;
 
+import java.util.InputMismatchException;
 import java.util.List;
 import java.util.Scanner;
 
@@ -37,7 +38,7 @@ public class BookingController extends Controller {
             case "delete" :
                 doDeleteBooking();
                 break;
-            case "reply" :
+            case "review" :
                 doWriteReview();
                 break;
             default :
@@ -281,7 +282,7 @@ public class BookingController extends Controller {
         }
 
         else {
-            System.out.printf("==== [%s]님 예약 현황 =======\n", loginedGuest.name);
+            System.out.printf("======== [%s]님 예약 현황 ========\n", loginedGuest.name);
             System.out.printf("** 오늘 날짜 : %s **\n", Util.getTodayDate());
             System.out.println("예약번호 | 객실 | 체크인날짜 | 체크아웃날짜");
 
@@ -292,7 +293,7 @@ public class BookingController extends Controller {
                 System.out.printf("    %4d |  %d | %4s |  %4s\n", bookingByGuest.id, bookingByGuest.roomId, bookingByGuest.checkInDate, bookingByGuest.checkOutDate);
             }
 
-            System.out.println("====================================");
+            System.out.println("======================================");
             System.out.print("예약을 취소하시겠습니까?) ");
             String answer = sc.nextLine();
 
@@ -301,20 +302,30 @@ public class BookingController extends Controller {
                 int answerId = sc.nextInt();
                 sc.nextLine();
 
-                // 예약번호에 따른 booking 가져오기
-                Booking forPrintBooking = bookingService.getForPrintBooking(answerId);
+                try {
+                    // 예약번호에 따른 booking 가져오기
+                    Booking forPrintBooking = bookingService.getForPrintBooking(answerId);
 
-                // 객실번호(303과 같은)를 100으로 나눴을 때 몫을 층 수
-                int floor = forPrintBooking.roomId / 100;
-                // 객실번호(303과 같은)를 100으로 나눴을 때 나머지를 호수
-                int number = forPrintBooking.roomId % 10;
+                    if(forPrintBooking.guestName.equals(loginedGuest.name)) {
+                        // 객실번호(303과 같은)를 100으로 나눴을 때 몫을 층 수
+                        int floor = forPrintBooking.roomId / 100;
+                        // 객실번호(303과 같은)를 100으로 나눴을 때 나머지를 호수
+                        int number = forPrintBooking.roomId % 10;
 
-                // 삭제할 forPrintBooking을 인자로 넘겨 room 정보 수정하기
-                Container.roomService.setBooingDelete(floor, number, forPrintBooking.checkInDate, forPrintBooking.checkOutDate);
-                // 예약목록에서 삭제
-                bookingService.deleteBooking(answerId);
+                        // 삭제할 forPrintBooking을 인자로 넘겨 room 정보 수정하기
+                        Container.roomService.setBooingDelete(floor, number, forPrintBooking.checkInDate, forPrintBooking.checkOutDate);
+                        // 예약목록에서 삭제
+                        bookingService.deleteBooking(answerId);
 
-                System.out.println("예약이 취소되었습니다!!");
+                        System.out.println("예약이 취소되었습니다!!");
+                    }
+                    // 만약 게스트 이름이 같지 않다면
+                    else {
+                        System.out.printf("입력하신 예약번호는 [%s]님의 예약 건이 아닙니다.\n", loginedGuest.name);
+                    }
+                } catch (NullPointerException | InputMismatchException e) {
+                    System.out.println("예약번호를 잘못 입력하셨습니다.");
+                }
             }
 
             else if(answer.equals("no")) {
@@ -328,5 +339,86 @@ public class BookingController extends Controller {
     }
 
     public void doWriteReview() {
+        // 로그인 된 게스트 가져오기
+        Guest loginedGuest = session.getLoginedGuest();
+
+        // 오늘 날짜로 체크아웃이 지난 booking 가져오기 (후기 작성한 예약목록 가져오기)
+        List<Booking> forBookingAbleReview = bookingService.getBookingsAbleReview(loginedGuest.name, Util.getTodayDate());
+
+        int foundBookingCount = forBookingAbleReview.size();
+        System.out.printf("[%s]님, 총 [%d]건에 대해 리뷰 작성이 가능합니다.\n", loginedGuest.name, foundBookingCount);
+
+        if(foundBookingCount == 0) {
+            System.out.println("리뷰 작성할 객실이 없습니다.");
+        }
+
+        else {
+            System.out.printf("======== [%s]님 숙박 이력 ========\n", loginedGuest.name);
+            System.out.printf("** 오늘 날짜 : %s **\n", Util.getTodayDate());
+            System.out.println("예약번호 | 객실 | 체크인날짜 | 체크아웃날짜");
+
+            // 예약한 목록 출력
+            for(int i = 0; i < forBookingAbleReview.size(); i++) {
+                Booking bookingByGuest = forBookingAbleReview.get(i);
+
+                System.out.printf("    %4d |  %d | %4s |  %4s\n", bookingByGuest.id, bookingByGuest.roomId, bookingByGuest.checkInDate, bookingByGuest.checkOutDate);
+            }
+
+            System.out.println("======================================");
+            System.out.print("리뷰를 작성하시겠습니까?) ");
+            String answer = sc.nextLine();
+
+            if(answer.equals("yes")) {
+                System.out.print("작성하고 싶은 예약번호 입력(숫자만) : ");
+                int answerId = sc.nextInt();
+                sc.nextLine();
+
+                try {
+                    // 예약번호에 따른 booking 가져오기
+                    Booking forPrintBooking = bookingService.getForPrintBooking(answerId);
+
+                    if(forPrintBooking.guestName.equals(loginedGuest.name)) {
+                        while(true) {
+                            System.out.println("** 평점은 총 5점 만점 중 소수점 첫째자리까지 입력 가능 **");
+                            System.out.print("평점 입력) ");
+                            double score = sc.nextDouble();
+                            sc.nextLine();
+
+                            // 소수 첫째자리 체크
+                            if(score != Math.round(score * 10) / 10.0) {
+                                System.out.println("평점은 소수점 첫째짜리까지만 입력이 가능합니다.");
+                                continue;
+                            }
+                            else if(score > 5) {
+                                System.out.println("평점은 최대 5점 만점입니다.");
+                                continue;
+                            }
+                            else {
+                                System.out.print("리뷰 작성) ");
+                                String reviewBody = sc.nextLine();
+
+                                Container.reviewService.doWrite(answerId, loginedGuest.id, reviewBody, score);
+                                System.out.println("리뷰가 등록되었습니다!!");
+                                break;
+                            }
+                        }
+                    }
+                    // 만약 게스트 이름이 같지 않다면
+                    else {
+                        System.out.printf("입력하신 예약번호는 [%s]님의 예약 건이 아닙니다.\n", loginedGuest.name);
+                    }
+                } catch (NullPointerException | InputMismatchException e) {
+                    System.out.println("예약번호를 잘못 입력하셨습니다.");
+                }
+            }
+
+            else if(answer.equals("no")) {
+                System.out.println("리뷰 작성을 중단합니다.");
+            }
+
+            else {
+                System.out.println("\'yes\' 또는 \'no\'를 입력해주세요");
+            }
+        }
     }
 }
